+++ src/main/groovy/org/nuisto/VersionBaseScriptClass.groovy
@@ -1,7 +1,0 @@
-package org.nuisto
-
-abstract class VersionBaseScriptClass extends Script {
-  void version(String version) {
-    this.binding.version = version
-  }
-}